pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-excel"
version = "0.3.0"
description = "Interactive CLI tool for adding decorated UPI QR codes to Excel rows."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "upi-qr-add maintainers" }
]
dependencies = [
  "typer[all]>=0.12.3",
  "rich>=13.7.1",
  "prompt_toolkit>=3.0.47",
  "openpyxl>=3.1.2",
  "qrcode[pil]>=7.4.2",
  "Pillow>=10.3.0",
  "tqdm>=4.66.4",
]

[project.urls]
Homepage = "https://github.com/ezhil-003/qr-excel"
Changelog = "https://github.com/ezhil-003/qr-excel/blob/main/upi_qr_add/CHANGELOG.md"
"Code of Conduct" = "https://github.com/ezhil-003/qr-excel/blob/main/upi_qr_add/CODE_OF_CONDUCT.md"

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
qr-excel = "qr_excel.cli.app:app"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["qr_excel*"]

[tool.setuptools.package-data]
qr_excel = ["assets/*.png", "assets/*.db"]

[tool.pytest.ini_options]
testpaths = ["tests"]
